added custom ingress rules to allow specific ingress depending on type of compute
