+++ pkg/infra/aws/modules/compute/macm1/macm1.go
@@ -4,12 +4,18 @@ import (
 	// "github.com/pulumi/pulumi-aws/sdk/v5/go/aws/elb"
 
 	"bytes"
+	"fmt"
 	"text/template"
 
+	"github.com/adrianriobo/qenvs/pkg/infra"
 	"github.com/adrianriobo/qenvs/pkg/infra/aws/services/ec2/ami"
 	"github.com/pulumi/pulumi-aws/sdk/v5/go/aws/ec2"
 	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
+
+	securityGroup "github.com/adrianriobo/qenvs/pkg/infra/aws/services/ec2/security-group"
 )
+
+const vncDefaultPort int = 5900
 
 func (r *MacM1Request) GetAMI(ctx *pulumi.Context) (*ec2.LookupAmiResult, error) {
 	return ami.GetAMIByName(ctx, r.Specs.AMI.RegexName, r.Specs.AMI.Owner, r.Specs.AMI.Filters)
@@ -23,6 +29,22 @@ func (r *MacM1Request) GetDedicatedHost(ctx *pulumi.Context) (*ec2.DedicatedHost
 			AvailabilityZone: pulumi.String(r.AvailabilityZones[0]),
 			InstanceType:     pulumi.String(r.Specs.InstaceTypes[0]),
 		})
+}
+
+func (r *MacM1Request) CustomIngressRules() []securityGroup.IngressRules {
+	return []securityGroup.IngressRules{
+		{
+			Description: fmt.Sprintf("VNC port for %s", r.Specs.ID),
+			FromPort:    vncDefaultPort,
+			ToPort:      vncDefaultPort,
+			Protocol:    "tcp",
+			CidrBlocks:  infra.NETWORKING_CIDR_ANY_IPV4,
+		},
+	}
+}
+
+func (r *MacM1Request) CustomSecurityGroups(ctx *pulumi.Context) ([]*ec2.SecurityGroup, error) {
+	return nil, nil
 }
 
 func (r *MacM1Request) GetPostScript() (string, error) {
