+++ pkg/infra/aws/modules/compute/rhel/rhel.go
@@ -3,6 +3,8 @@ package rhel
 import (
 	"fmt"
 
+	securityGroup "github.com/adrianriobo/qenvs/pkg/infra/aws/services/ec2/security-group"
+
 	"github.com/adrianriobo/qenvs/pkg/infra/aws/services/ec2/ami"
 	"github.com/pulumi/pulumi-aws/sdk/v5/go/aws/ec2"
 	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
@@ -14,6 +16,14 @@ func (r *RHELRequest) GetAMI(ctx *pulumi.Context) (*ec2.LookupAmiResult, error) 
 }
 
 func (r *RHELRequest) GetDedicatedHost(ctx *pulumi.Context) (*ec2.DedicatedHost, error) {
+	return nil, nil
+}
+
+func (r *RHELRequest) CustomIngressRules() []securityGroup.IngressRules {
+	return nil
+}
+
+func (r *RHELRequest) CustomSecurityGroups(ctx *pulumi.Context) ([]*ec2.SecurityGroup, error) {
 	return nil, nil
 }
